refactor: move source to a core folder
