+++ src/app/home/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import MakeHome from '@/main/factories/app/home/page';
+import MakeHome from '@/core/main/factories/app/home/page';
 
 export default function Page() {
   return <MakeHome />;
