+++ src/app/providers.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import '@/application/global.css';
+import '@/core/application/global.css';
 
 import { Analytics } from '@vercel/analytics/react';
 import { SpeedInsights } from '@vercel/speed-insights/next';
