+++ src/core/application/components/layout/footer.tsx
