+++ src/core/application/global.css
