+++ src/core/application/pages/home/components/timer.tsx
