+++ src/core/application/pages/home/components/wpm.tsx
