+++ src/core/infrastructure/services/useWikipedia.ts
@@ -1,4 +1,4 @@
-import { htmlToText, symbolsToText } from '@/infrastructure/utils';
+import { htmlToText, symbolsToText } from '@/core/infrastructure/utils';
 import { useQuery } from '@tanstack/react-query';
 
 export type WikipediaRandomWords = string[];
