+++ src/core/infrastructure/utils.ts
