+++ src/core/main/factories/app/home/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { HomePage } from '@/application/pages/home';
-import { UseWikipedia } from '@/infrastructure/services/useWikipedia';
+import { HomePage } from '@/core/application/pages/home';
+import { UseWikipedia } from '@/core/infrastructure/services/useWikipedia';
 
 const MakeHome = () => {
   const { query } = UseWikipedia();
