+++ tests/__unit__/app/application/components/layout/footer.spec.ts
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 
 import { render } from '@testing-library/react';
 
-import { Footer } from '@/application/components/layout/footer';
+import { Footer } from '@/core/application/components/layout/footer';
 
 describe('Footer Component', () => {
   test('Should renders correctly', () => {
