+++ tests/__unit__/app/application/components/layout/header.spec.ts
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 
 import { render } from '@testing-library/react';
 
-import { Header } from '@/application/components/layout/header';
+import { Header } from '@/core/application/components/layout/header';
 
 describe('Header Component', () => {
   test('Should renders correctly', () => {
