+++ tsconfig.json
@@ -23,7 +23,7 @@
     ],
     "paths": {
       "@/*": [
-        "./src/app/*"
+        "./src/*"
       ]
     }
   }
